// types/contact.types.ts
export interface ContactDto {
  firstName: string
  lastName: string
  fullName: string
  email: string
  email2: string
  phone: string
  phone2: string
  whatsapp: string
  company: string
  jobTitle: string
  website: string
  address: string
  city: string
  state: string
  country: string
  zipCode: string
  linkedin: string
  twitter: string
  facebook: string
  instagram: string
  status: string
  tags: string[]
  notes: string
}